+++ packages/wonder-blocks-accordion/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @khanacademy/wonder-blocks-accordion
 
+## 3.1.56
+
+### Patch Changes
+
+- Updated dependencies [40f217b]
+    - @khanacademy/wonder-blocks-tokens@16.3.0
+    - @khanacademy/wonder-blocks-clickable@8.1.7
+    - @khanacademy/wonder-blocks-icon@5.3.11
+    - @khanacademy/wonder-blocks-typography@4.3.1
+
+## 3.1.55
+
+### Patch Changes
+
+- 3bba434: Fixed and unskipped tests
+
 ## 3.1.54
 
 ### Patch Changes