+++ packages/wonder-blocks-form/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @khanacademy/wonder-blocks-form
 
+## 7.5.8
+
+### Patch Changes
+
+- Updated dependencies [40f217b]
+    - @khanacademy/wonder-blocks-tokens@16.3.0
+    - @khanacademy/wonder-blocks-clickable@8.1.7
+    - @khanacademy/wonder-blocks-icon@5.3.11
+    - @khanacademy/wonder-blocks-layout@3.1.48
+    - @khanacademy/wonder-blocks-typography@4.3.1
+
 ## 7.5.7
 
 ### Patch Changes